// src/clients/generated/commonApi/models/createContractRequest.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */
import type { CreateContractRequestStatus } from './createContractRequestStatus';

/**
 * CreateContractRequest
 */
export interface CreateContractRequest {
  /**
   * Дата договора
   * @maxLength 10
   */
  contractDate?: string;
  /**
   * Номер договора
   * @maxLength 255
   */
  contractNumber: string;
  /**
   * Идентификатор контрагента
   * @maxLength 36
   */
  counterpartyId: string;
  /**
   * Дата окончания действия договора
   * @maxLength 10
   */
  endDate?: string;
  /**
   * Примечания к договору
   * @maxLength 255
   */
  notes?: string;
  /**
   * Дата, с которой договор вступает в сил
   * @maxLength 10
   */
  startDate?: string;
  /**
   * Статус договора
   * @maxLength 255
   */
  status: CreateContractRequestStatus;
}