// src/clients/generated/commonApi/models/createContractRequestStatus.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */

/**
 * Статус договора
 * @maxLength 255
 */
export type CreateContractRequestStatus = typeof CreateContractRequestStatus[keyof typeof CreateContractRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateContractRequestStatus = {
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  SUSPENDED: 'SUSPENDED',
  TERMINATED: 'TERMINATED',
} as const;